Chatrooms refactor

+++ backend/src/apps/ws_authentication/services.py
@@ -1,4 +1,4 @@
-from channels.exceptions import DenyConnection
+from django.core.exceptions import ValidationError, ObjectDoesNotExist
 
 from src.apps.authentication.models import User
 from src.apps.ws_authentication.models import Ticket
@@ -12,13 +12,15 @@ def ws_authenticate(*, token: str, ip_address) -> User:
     try:
         ticket = Ticket.objects.select_related("user").get(token=token)
     except Ticket.DoesNotExist:
-        raise DenyConnection()
+        raise ObjectDoesNotExist()
 
     if ticket.is_expired or ticket.ip_address != ip_address:
         ticket.delete()
-        raise DenyConnection()
+        raise ValidationError("Ticket has expired")
+
+    if ticket.ip_address != ip_address:
+        raise ValidationError("Client and ticket ip addresses do not match.")
 
     ticket.delete()
 
     return ticket.user
-
